Update configuration and game simulator for training adjustments and reward recalibration
- Increased `BUFFER_SIZE` in `config.py` to enhance replay memory capacity.
- Reduced `N_STEPS` in `config.py` for more frequent network updates.
- Changed data type from `float16` to `float32` in `dqn_agent.py` for improved precision.
- Set `cluster_reward` to 0 and adjusted `step_reward` calculations in `game_simulator.py` for better reward dynamics.
- Updated win reward calculation in `game_simulator.py` for improved game balance.

+++ ai_training/dqn_agent.py
@@ -117,12 +117,12 @@ class DQNAgent:
 
             self.memory.append(
                 (
-                    first_state[0].astype(np.float16),  # board
-                    first_state[1].astype(np.float16),  # global features
+                    first_state[0].astype(np.float32),  # board
+                    first_state[1].astype(np.float32),  # global features
                     swap_to_idx(*first_action[0], *first_action[1]),
                     R,
-                    next_state[0].astype(np.float16),   # board n‑steps ahead
-                    next_state[1].astype(np.float16),   # global features n‑steps ahead
+                    next_state[0].astype(np.float32),   # board n‑steps ahead
+                    next_state[1].astype(np.float32),   # global features n‑steps ahead
                     done,
                     [swap_to_idx(*s[0], *s[1]) for s in valid_next],
                     effective_n,
